@@ -23,8 +23,8 @@ exclude = [
 ]
 build = "build.rs"
 autotests = false
-edition = "2018"
-rust-version = "1.70"
+edition = "2021"
+rust-version = "1.72"
 
 [[bin]]
 bench = false
@@ -49,12 +49,11 @@ members = [
 ]
 
 [dependencies]
+anyhow = "1.0.75"
 bstr = "1.6.0"
 grep = { version = "0.2.12", path = "crates/grep" }
 ignore = { version = "0.4.19", path = "crates/ignore" }
-lazy_static = "1.1.0"
 log = "0.4.5"
-regex = { version = "1.1", default-features = false }
 serde_json = "1.0.23"
 termcolor = "1.1.0"
 
@@ -63,11 +62,13 @@ version = "2.33.0"
 default-features = false
 features = ["suggestions"]
 
+[dependencies.regex]
+version = "1.3.0"
+default-features = false
+features = ["std"]
+
 [target.'cfg(all(target_env = "musl", target_pointer_width = "64"))'.dependencies.jemallocator]
 version = "0.5.0"
-
-[build-dependencies]
-lazy_static = "1.1.0"
 
 [build-dependencies.clap]
 version = "2.33.0"
