@@ -11,16 +11,16 @@ repository = "https://github.com/BurntSushi/ripgrep/tree/master/crates/cli"
 readme = "README.md"
 keywords = ["regex", "grep", "cli", "utility", "util"]
 license = "Unlicense OR MIT"
-edition = "2018"
+edition = "2021"
 
 [dependencies]
-bstr = "1.6.0"
+bstr = { version = "1.6.2", features = ["std"] }
 globset = { version = "0.4.10", path = "../globset" }
-lazy_static = "1.1.0"
-log = "0.4.5"
-regex = "1.1"
-same-file = "1.0.4"
-termcolor = "1.0.4"
+log = "0.4.20"
+termcolor = "1.3.0"
 
 [target.'cfg(windows)'.dependencies.winapi-util]
-version = "0.1.1"
+version = "0.1.6"
+
+[target.'cfg(unix)'.dependencies.libc]
+version = "0.2.148"
