@@ -12,27 +12,38 @@ repository = "https://github.com/BurntSushi/ripgrep/tree/master/crates/ignore"
 readme = "README.md"
 keywords = ["glob", "ignore", "gitignore", "pattern", "file"]
 license = "Unlicense OR MIT"
-edition = "2018"
+edition = "2021"
 
 [lib]
 name = "ignore"
 bench = false
 
 [dependencies]
+crossbeam-deque = "0.8.3"
 globset = { version = "0.4.10", path = "../globset" }
-lazy_static = "1.1"
-log = "0.4.5"
-memchr = "2.5"
-regex = { version = "1.9.0", default-features = false, features = ["perf", "std", "unicode-gencat"] }
-same-file = "1.0.4"
-thread_local = "1"
-walkdir = "2.2.7"
+log = "0.4.20"
+memchr = "2.6.3"
+same-file = "1.0.6"
+walkdir = "2.4.0"
+
+[dependencies.regex]
+version = "1.3.0"
+default-features = false
+features = ["std"]
+
+[dependencies.regex-automata]
+version = "0.4.0"
+default-features = false
+features = ["std", "perf", "syntax", "meta", "nfa", "hybrid", "dfa-onepass"]
 
 [target.'cfg(windows)'.dependencies.winapi-util]
 version = "0.1.2"
 
 [dev-dependencies]
-crossbeam-channel = "0.5.0"
+bstr = { version = "1.6.2", default-features = false, features = ["std"] }
+crossbeam-channel = "0.5.8"
 
 [features]
-simd-accel = ["globset/simd-accel"]
+# DEPRECATED. It is a no-op. SIMD is done automatically through runtime
+# dispatch.
+simd-accel = []
