@@ -10,7 +10,6 @@
 // it into a ripgrep-specific configuration type that is not coupled with clap.
 
 use clap::{self, crate_authors, crate_version, App, AppSettings};
-use lazy_static::lazy_static;
 
 const ABOUT: &str = "
 ripgrep (rg) recursively searches the current directory for a regex pattern.
@@ -47,18 +46,19 @@ OPTIONS:
 
 /// Build a clap application parameterized by usage strings.
 pub fn app() -> App<'static, 'static> {
+    use std::sync::OnceLock;
+
     // We need to specify our version in a static because we've painted clap
     // into a corner. We've told it that every string we give it will be
     // 'static, but we need to build the version string dynamically. We can
     // fake the 'static lifetime with lazy_static.
-    lazy_static! {
-        static ref LONG_VERSION: String = long_version(None, true);
-    }
+    static LONG_VERSION: OnceLock<String> = OnceLock::new();
+    let long_version = LONG_VERSION.get_or_init(|| long_version(None, true));
 
     let mut app = App::new("ripgrep")
         .author(crate_authors!())
         .version(crate_version!())
-        .long_version(LONG_VERSION.as_str())
+        .long_version(long_version.as_str())
         .about(ABOUT)
         .max_term_width(100)
         .setting(AppSettings::UnifiedHelpMessage)
@@ -584,6 +584,8 @@ pub fn all_args_and_flags() -> Vec<RGArg> {
     flag_glob_case_insensitive(&mut args);
     flag_heading(&mut args);
     flag_hidden(&mut args);
+    flag_hostname_bin(&mut args);
+    flag_hyperlink_format(&mut args);
     flag_iglob(&mut args);
     flag_ignore_case(&mut args);
     flag_ignore_file(&mut args);
@@ -1510,6 +1512,113 @@ This flag can be disabled with --no-hidden.
     args.push(arg);
 
     let arg = RGArg::switch("no-hidden").hidden().overrides("hidden");
+    args.push(arg);
+}
+
+fn flag_hostname_bin(args: &mut Vec<RGArg>) {
+    const SHORT: &str = "Run a program to get this system's hostname.";
+    const LONG: &str = long!(
+        "\
+This flag controls how ripgrep determines this system's hostname. The flag's
+value should correspond to an executable (either a path or something that can
+be found via your system's *PATH* environment variable). When set, ripgrep will
+run this executable, with no arguments, and treat its output (with leading and
+trailing whitespace stripped) as your system's hostname.
+
+When not set (the default, or the empty string), ripgrep will try to
+automatically detect your system's hostname. On Unix, this corresponds
+to calling *gethostname*. On Windows, this corresponds to calling
+*GetComputerNameExW* to fetch the system's \"physical DNS hostname.\"
+
+ripgrep uses your system's hostname for producing hyperlinks.
+"
+    );
+    let arg =
+        RGArg::flag("hostname-bin", "COMMAND").help(SHORT).long_help(LONG);
+    args.push(arg);
+}
+
+fn flag_hyperlink_format(args: &mut Vec<RGArg>) {
+    const SHORT: &str = "Set the format of hyperlinks to match results.";
+    const LONG: &str = long!(
+        "\
+Set the format of hyperlinks to match results. Hyperlinks make certain elements
+of ripgrep's output, such as file paths, clickable. This generally only works
+in terminal emulators that support OSC-8 hyperlinks. For example, the format
+file://{host}{path} will emit an RFC 8089 hyperlink. To see the format that
+ripgrep is using, pass the --debug flag.
+
+Alternatively, a format string may correspond to one of the following aliases:
+default, file, grep+, kitty, macvim, none, textmate, vscode, vscode-insiders,
+vscodium. The alias will be replaced with a format string that is intended to
+work for the corresponding application.
+
+The following variables are available in the format string:
+
+{path}: Required. This is replaced with a path to a matching file. The
+path is guaranteed to be absolute and percent encoded such that it is valid to
+put into a URI. Note that a path is guaranteed to start with a /.
+
+{host}: Optional. This is replaced with your system's hostname. On Unix,
+this corresponds to calling 'gethostname'. On Windows, this corresponds to
+calling 'GetComputerNameExW' to fetch the system's \"physical DNS hostname.\"
+Alternatively, if --hostname-bin was provided, then the hostname returned from
+the output of that program will be returned. If no hostname could be found,
+then this variable is replaced with the empty string.
+
+{line}: Optional. If appropriate, this is replaced with the line number of
+a match. If no line number is available (for example, if --no-line-number was
+given), then it is automatically replaced with the value 1.
+
+{column}: Optional, but requires the presence of {line}. If appropriate, this
+is replaced with the column number of a match. If no column number is available
+(for example, if --no-column was given), then it is automatically replaced with
+the value 1.
+
+{wslprefix}: Optional. This is a special value that is set to
+wsl$/WSL_DISTRO_NAME, where WSL_DISTRO_NAME corresponds to the value of
+the equivalent environment variable. If the system is not Unix or if the
+WSL_DISTRO_NAME environment variable is not set, then this is replaced with the
+empty string.
+
+A format string may be empty. An empty format string is equivalent to the
+'none' alias. In this case, hyperlinks will be disabled.
+
+At present, ripgrep does not enable hyperlinks by default. Users must opt into
+them. If you aren't sure what format to use, try 'default'.
+
+When ripgrep detects a tty on stdout then hyperlinks are automatically
+disabled, regardless of the value of this flag. Users can pass '--color always'
+to forcefully emit hyperlinks.
+
+Note that hyperlinks are only written when a path is also in the output
+and colors are enabled. To write hyperlinks without colors, you'll need to
+configure ripgrep to not colorize anything without actually disabling all ANSI
+escape codes completely:
+
+    --colors 'path:none' --colors 'line:none' --colors 'column:none' --colors 'match:none'
+
+ripgrep works this way because it treats the --color=(never|always|auto) flag
+as a proxy for whether ANSI escape codes should be used at all. This means
+that environment variables like NO_COLOR=1 and TERM=dumb not only disable
+colors, but hyperlinks as well. Similarly, colors and hyperlinks are disabled
+when ripgrep is not writing to a tty. (Unless one forces the issue by setting
+--color=always.)
+
+If you're searching a file directly, for example:
+
+    rg foo path/to/file
+
+then hyperlinks will not be emitted since the path given does not appear
+in the output. To make the path appear, and thus also a hyperlink, use the
+-H/--with-filename flag.
+
+For more information on hyperlinks in terminal emulators, see:
+https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
+"
+    );
+    let arg =
+        RGArg::flag("hyperlink-format", "FORMAT").help(SHORT).long_help(LONG);
     args.push(arg);
 }
 
