@@ -2,17 +2,12 @@
 // primary output of these routines is a sequence of arguments, where each
 // argument corresponds precisely to one shell argument.
 
-use std::env;
-use std::error::Error;
-use std::ffi::OsString;
-use std::fs::File;
-use std::io;
-use std::path::{Path, PathBuf};
+use std::{
+    ffi::OsString,
+    path::{Path, PathBuf},
+};
 
 use bstr::{io::BufReadExt, ByteSlice};
-use log;
-
-use crate::Result;
 
 /// On Unix, /etc/ripgreprc will be used as a default config file if RIPGREP_CONFIG_PATH is unset.
 #[cfg(unix)]
@@ -25,7 +20,7 @@ const DEFAULT_CONFIG_PATH: Option<&str> = None;
 /// Return a sequence of arguments derived from ripgrep rc configuration files.
 pub fn args() -> Vec<OsString> {
     let (config_path, used_default) =
-        match (env::var_os("RIPGREP_CONFIG_PATH"), DEFAULT_CONFIG_PATH) {
+        match (std::env::var_os("RIPGREP_CONFIG_PATH"), DEFAULT_CONFIG_PATH) {
             (Some(config_path), _) => {
                 if config_path.is_empty() {
                     return vec![];
@@ -71,11 +66,11 @@ pub fn args() -> Vec<OsString> {
 /// for each line in addition to successfully parsed arguments.
 fn parse<P: AsRef<Path>>(
     path: P,
-) -> Result<(Vec<OsString>, Vec<Box<dyn Error>>)> {
+) -> anyhow::Result<(Vec<OsString>, Vec<anyhow::Error>)> {
     let path = path.as_ref();
-    match File::open(&path) {
+    match std::fs::File::open(&path) {
         Ok(file) => parse_reader(file),
-        Err(err) => Err(From::from(format!("{}: {}", path.display(), err))),
+        Err(err) => anyhow::bail!("{}: {}", path.display(), err),
     }
 }
 
@@ -90,10 +85,10 @@ fn parse<P: AsRef<Path>>(
 /// If the reader could not be read, then an error is returned. If there was a
 /// problem parsing one or more lines, then errors are returned for each line
 /// in addition to successfully parsed arguments.
-fn parse_reader<R: io::Read>(
+fn parse_reader<R: std::io::Read>(
     rdr: R,
-) -> Result<(Vec<OsString>, Vec<Box<dyn Error>>)> {
-    let mut bufrdr = io::BufReader::new(rdr);
+) -> anyhow::Result<(Vec<OsString>, Vec<anyhow::Error>)> {
+    let mut bufrdr = std::io::BufReader::new(rdr);
     let (mut args, mut errs) = (vec![], vec![]);
     let mut line_number = 0;
     bufrdr.for_byte_line_with_terminator(|line| {
@@ -108,7 +103,7 @@ fn parse_reader<R: io::Read>(
                 args.push(osstr.to_os_string());
             }
             Err(err) => {
-                errs.push(format!("{}: {}", line_number, err).into());
+                errs.push(anyhow::anyhow!("{line_number}: {err}"));
             }
         }
         Ok(true)
