@@ -1,24 +1,26 @@
-use std::cmp;
-use std::ffi::OsStr;
-use std::fmt;
-use std::fs::{self, FileType, Metadata};
-use std::io;
-use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-use std::sync::{Arc, Mutex};
-use std::thread;
-use std::time::Duration;
-use std::vec;
+use std::{
+    cmp::Ordering,
+    ffi::OsStr,
+    fs::{self, FileType, Metadata},
+    io,
+    path::{Path, PathBuf},
+    sync::atomic::{AtomicBool, AtomicUsize, Ordering as AtomicOrdering},
+    sync::Arc,
+};
 
-use regex::Regex;
-use same_file::Handle;
-use walkdir::{self, WalkDir};
+use {
+    crossbeam_deque::{Stealer, Worker as Deque},
+    same_file::Handle,
+    walkdir::{self, WalkDir},
+};
 
-use crate::dir::{Ignore, IgnoreBuilder};
-use crate::gitignore::GitignoreBuilder;
-use crate::overrides::Override;
-use crate::types::Types;
-use crate::{Error, PartialErrorBuilder};
+use crate::{
+    dir::{Ignore, IgnoreBuilder},
+    gitignore::GitignoreBuilder,
+    overrides::Override,
+    types::Types,
+    Error, PartialErrorBuilder,
+};
 
 /// A directory entry with a possible error attached.
 ///
@@ -37,9 +39,7 @@ impl DirEntry {
     }
 
     /// The full path that this entry represents.
-    /// Analogous to [`path`], but moves ownership of the path.
-    ///
-    /// [`path`]: struct.DirEntry.html#method.path
+    /// Analogous to [`DirEntry::path`], but moves ownership of the path.
     pub fn into_path(self) -> PathBuf {
         self.dent.into_path()
     }
@@ -108,11 +108,11 @@ impl DirEntry {
     }
 
     fn new_walkdir(dent: walkdir::DirEntry, err: Option<Error>) -> DirEntry {
-        DirEntry { dent: DirEntryInner::Walkdir(dent), err: err }
+        DirEntry { dent: DirEntryInner::Walkdir(dent), err }
     }
 
     fn new_raw(dent: DirEntryRaw, err: Option<Error>) -> DirEntry {
-        DirEntry { dent: DirEntryInner::Raw(dent), err: err }
+        DirEntry { dent: DirEntryInner::Raw(dent), err }
     }
 }
 
@@ -252,8 +252,8 @@ struct DirEntryRaw {
     metadata: fs::Metadata,
 }
 
-impl fmt::Debug for DirEntryRaw {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+impl std::fmt::Debug for DirEntryRaw {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         // Leaving out FileType because it doesn't have a debug impl
         // in Rust 1.9. We could add it if we really wanted to by manually
         // querying each possibly file type. Meh. ---AG
@@ -325,7 +325,7 @@ impl DirEntryRaw {
     ) -> Result<DirEntryRaw, Error> {
         let ty = ent.file_type().map_err(|err| {
             let err = Error::Io(io::Error::from(err)).with_path(ent.path());
-            Error::WithDepth { depth: depth, err: Box::new(err) }
+            Error::WithDepth { depth, err: Box::new(err) }
         })?;
         DirEntryRaw::from_entry_os(depth, ent, ty)
     }
@@ -338,13 +338,13 @@ impl DirEntryRaw {
     ) -> Result<DirEntryRaw, Error> {
         let md = ent.metadata().map_err(|err| {
             let err = Error::Io(io::Error::from(err)).with_path(ent.path());
-            Error::WithDepth { depth: depth, err: Box::new(err) }
+            Error::WithDepth { depth, err: Box::new(err) }
         })?;
         Ok(DirEntryRaw {
             path: ent.path(),
-            ty: ty,
+            ty,
             follow_link: false,
-            depth: depth,
+            depth,
             metadata: md,
         })
     }
@@ -359,9 +359,9 @@ impl DirEntryRaw {
 
         Ok(DirEntryRaw {
             path: ent.path(),
-            ty: ty,
+            ty,
             follow_link: false,
-            depth: depth,
+            depth,
             ino: ent.ino(),
         })
     }
@@ -392,7 +392,7 @@ impl DirEntryRaw {
             path: pb,
             ty: md.file_type(),
             follow_link: link,
-            depth: depth,
+            depth,
             metadata: md,
         })
     }
@@ -411,7 +411,7 @@ impl DirEntryRaw {
             path: pb,
             ty: md.file_type(),
             follow_link: link,
-            depth: depth,
+            depth,
             ino: md.ino(),
         })
     }
@@ -495,17 +495,15 @@ pub struct WalkBuilder {
 
 #[derive(Clone)]
 enum Sorter {
-    ByName(
-        Arc<dyn Fn(&OsStr, &OsStr) -> cmp::Ordering + Send + Sync + 'static>,
-    ),
-    ByPath(Arc<dyn Fn(&Path, &Path) -> cmp::Ordering + Send + Sync + 'static>),
+    ByName(Arc<dyn Fn(&OsStr, &OsStr) -> Ordering + Send + Sync + 'static>),
+    ByPath(Arc<dyn Fn(&Path, &Path) -> Ordering + Send + Sync + 'static>),
 }
 
 #[derive(Clone)]
 struct Filter(Arc<dyn Fn(&DirEntry) -> bool + Send + Sync + 'static>);
 
-impl fmt::Debug for WalkBuilder {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+impl std::fmt::Debug for WalkBuilder {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("WalkBuilder")
             .field("paths", &self.paths)
             .field("ig_builder", &self.ig_builder)
@@ -579,7 +577,7 @@ impl WalkBuilder {
             .into_iter();
         let ig_root = self.ig_builder.build();
         Walk {
-            its: its,
+            its,
             it: None,
             ig_root: ig_root.clone(),
             ig: ig_root.clone(),
@@ -691,7 +689,7 @@ impl WalkBuilder {
     /// This is an additional whitelist applied after all other logic
     pub fn file_search_regex(
         &mut self,
-        re: Option<Regex>,
+        re: Option<regex::Regex>,
     ) -> &mut WalkBuilder {
         self.ig_builder.file_search_regex(re);
         self
@@ -840,7 +838,7 @@ impl WalkBuilder {
     /// Note that this is not used in the parallel iterator.
     pub fn sort_by_file_path<F>(&mut self, cmp: F) -> &mut WalkBuilder
     where
-        F: Fn(&Path, &Path) -> cmp::Ordering + Send + Sync + 'static,
+        F: Fn(&Path, &Path) -> Ordering + Send + Sync + 'static,
     {
         self.sorter = Some(Sorter::ByPath(Arc::new(cmp)));
         self
@@ -859,7 +857,7 @@ impl WalkBuilder {
     /// Note that this is not used in the parallel iterator.
     pub fn sort_by_file_name<F>(&mut self, cmp: F) -> &mut WalkBuilder
     where
-        F: Fn(&OsStr, &OsStr) -> cmp::Ordering + Send + Sync + 'static,
+        F: Fn(&OsStr, &OsStr) -> Ordering + Send + Sync + 'static,
     {
         self.sorter = Some(Sorter::ByName(Arc::new(cmp)));
         self
@@ -923,7 +921,7 @@ impl WalkBuilder {
 /// ignore files like `.gitignore` are respected. The precise matching rules
 /// and precedence is explained in the documentation for `WalkBuilder`.
 pub struct Walk {
-    its: vec::IntoIter<(PathBuf, Option<WalkEventIter>)>,
+    its: std::vec::IntoIter<(PathBuf, Option<WalkEventIter>)>,
     it: Option<WalkEventIter>,
     ig_root: Ignore,
     ig: Ignore,
@@ -1052,6 +1050,8 @@ impl Iterator for Walk {
     }
 }
 
+impl std::iter::FusedIterator for Walk {}
+
 /// WalkEventIter transforms a WalkDir iterator into an iterator that more
 /// accurately describes the directory tree. Namely, it emits events that are
 /// one of three types: directory, file or "exit." An "exit" event means that
@@ -1135,10 +1135,10 @@ impl WalkState {
     }
 }
 
-/// A builder for constructing a visitor when using
-/// [`WalkParallel::visit`](struct.WalkParallel.html#method.visit). The builder
-/// will be called for each thread started by `WalkParallel`. The visitor
-/// returned from each builder is then called for every directory entry.
+/// A builder for constructing a visitor when using [`WalkParallel::visit`].
+/// The builder will be called for each thread started by `WalkParallel`. The
+/// visitor returned from each builder is then called for every directory
+/// entry.
 pub trait ParallelVisitorBuilder<'s> {
     /// Create per-thread `ParallelVisitor`s for `WalkParallel`.
     fn build(&mut self) -> Box<dyn ParallelVisitor + 's>;
@@ -1155,9 +1155,8 @@ impl<'a, 's, P: ParallelVisitorBuilder<'s>> ParallelVisitorBuilder<'s>
 /// Receives files and directories for the current thread.
 ///
 /// Setup for the traversal can be implemented as part of
-/// [`ParallelVisitorBuilder::build`](trait.ParallelVisitorBuilder.html#tymethod.build).
-/// Teardown when traversal finishes can be implemented by implementing the
-/// `Drop` trait on your traversal type.
+/// [`ParallelVisitorBuilder::build`]. Teardown when traversal finishes can be
+/// implemented by implementing the `Drop` trait on your traversal type.
 pub trait ParallelVisitor: Send {
     /// Receives files and directories for the current thread. This is called
     /// once for every directory entry visited by traversal.
@@ -1199,7 +1198,7 @@ impl<'s> ParallelVisitor for FnVisitorImp<'s> {
 ///
 /// Unlike `Walk`, this uses multiple threads for traversing a directory.
 pub struct WalkParallel {
-    paths: vec::IntoIter<PathBuf>,
+    paths: std::vec::IntoIter<PathBuf>,
     ig_root: Ignore,
     max_filesize: Option<u64>,
     max_depth: Option<usize>,
@@ -1240,9 +1239,8 @@ impl WalkParallel {
     /// can be merged together into a single data structure.
     pub fn visit(mut self, builder: &mut dyn ParallelVisitorBuilder<'_>) {
         let threads = self.threads();
-        let stack = Arc::new(Mutex::new(vec![]));
+        let mut stack = vec![];
         {
-            let mut stack = stack.lock().unwrap();
             let mut visitor = builder.build();
             let mut paths = Vec::new().into_iter();
             std::mem::swap(&mut paths, &mut self.paths);
@@ -1280,9 +1278,9 @@ impl WalkParallel {
                     }
                 };
                 stack.push(Message::Work(Work {
-                    dent: dent,
+                    dent,
                     ignore: self.ig_root.clone(),
-                    root_device: root_device,
+                    root_device,
                 }));
             }
             // ... but there's no need to start workers if we don't need them.
@@ -1292,14 +1290,14 @@ impl WalkParallel {
         }
         // Create the workers and then wait for them to finish.
         let quit_now = Arc::new(AtomicBool::new(false));
-        let num_pending =
-            Arc::new(AtomicUsize::new(stack.lock().unwrap().len()));
+        let num_pending = Arc::new(AtomicUsize::new(stack.len()));
+        let stacks = Stack::new_for_each_thread(threads, stack);
         std::thread::scope(|s| {
-            let mut handles = vec![];
-            for _ in 0..threads {
-                let worker = Worker {
+            let handles: Vec<_> = stacks
+                .into_iter()
+                .map(|stack| Worker {
                     visitor: builder.build(),
-                    stack: stack.clone(),
+                    stack,
                     quit_now: quit_now.clone(),
                     num_pending: num_pending.clone(),
                     max_depth: self.max_depth,
@@ -1307,9 +1305,9 @@ impl WalkParallel {
                     follow_links: self.follow_links,
                     skip: self.skip.clone(),
                     filter: self.filter.clone(),
-                };
-                handles.push(s.spawn(|| worker.run()));
-            }
+                })
+                .map(|worker| s.spawn(|| worker.run()))
+                .collect();
             for handle in handles {
                 handle.join().unwrap();
             }
@@ -1399,6 +1397,73 @@ impl Work {
     }
 }
 
+/// A work-stealing stack.
+#[derive(Debug)]
+struct Stack {
+    /// This thread's index.
+    index: usize,
+    /// The thread-local stack.
+    deque: Deque<Message>,
+    /// The work stealers.
+    stealers: Arc<[Stealer<Message>]>,
+}
+
+impl Stack {
+    /// Create a work-stealing stack for each thread. The given messages
+    /// correspond to the initial paths to start the search at. They will
+    /// be distributed automatically to each stack in a round-robin fashion.
+    fn new_for_each_thread(threads: usize, init: Vec<Message>) -> Vec<Stack> {
+        // Using new_lifo() ensures each worker operates depth-first, not
+        // breadth-first. We do depth-first because a breadth first traversal
+        // on wide directories with a lot of gitignores is disastrous (for
+        // example, searching a directory tree containing all of crates.io).
+        let deques: Vec<Deque<Message>> =
+            std::iter::repeat_with(Deque::new_lifo).take(threads).collect();
+        let stealers = Arc::<[Stealer<Message>]>::from(
+            deques.iter().map(Deque::stealer).collect::<Vec<_>>(),
+        );
+        let stacks: Vec<Stack> = deques
+            .into_iter()
+            .enumerate()
+            .map(|(index, deque)| Stack {
+                index,
+                deque,
+                stealers: stealers.clone(),
+            })
+            .collect();
+        // Distribute the initial messages.
+        init.into_iter()
+            .zip(stacks.iter().cycle())
+            .for_each(|(m, s)| s.push(m));
+        stacks
+    }
+
+    /// Push a message.
+    fn push(&self, msg: Message) {
+        self.deque.push(msg);
+    }
+
+    /// Pop a message.
+    fn pop(&self) -> Option<Message> {
+        self.deque.pop().or_else(|| self.steal())
+    }
+
+    /// Steal a message from another queue.
+    fn steal(&self) -> Option<Message> {
+        // For fairness, try to steal from index + 1, index + 2, ... len - 1,
+        // then wrap around to 0, 1, ... index - 1.
+        let (left, right) = self.stealers.split_at(self.index);
+        // Don't steal from ourselves
+        let right = &right[1..];
+
+        right
+            .iter()
+            .chain(left.iter())
+            .map(|s| s.steal_batch_and_pop(&self.deque))
+            .find_map(|s| s.success())
+    }
+}
+
 /// A worker is responsible for descending into directories, updating the
 /// ignore matchers, producing new work and invoking the caller's callback.
 ///
@@ -1406,13 +1471,13 @@ impl Work {
 struct Worker<'s> {
     /// The caller's callback.
     visitor: Box<dyn ParallelVisitor + 's>,
-    /// A stack of work to do.
+    /// A work-stealing stack of work to do.
     ///
     /// We use a stack instead of a channel because a stack lets us visit
     /// directories in depth first order. This can substantially reduce peak
-    /// memory usage by keeping both the number of files path and gitignore
+    /// memory usage by keeping both the number of file paths and gitignore
     /// matchers in memory lower.
-    stack: Arc<Mutex<Vec<Message>>>,
+    stack: Stack,
     /// Whether all workers should terminate at the next opportunity. Note
     /// that we need this because we don't want other `Work` to be done after
     /// we quit. We wouldn't need this if have a priority channel.
@@ -1652,7 +1717,8 @@ impl<'s> Worker<'s> {
                         // CPU waiting, we let the thread sleep for a bit. In
                         // general, this tends to only occur once the search is
                         // approaching termination.
-                        thread::sleep(Duration::from_millis(1));
+                        let dur = std::time::Duration::from_millis(1);
+                        std::thread::sleep(dur);
                     }
                 }
             }
@@ -1661,41 +1727,38 @@ impl<'s> Worker<'s> {
 
     /// Indicates that all workers should quit immediately.
     fn quit_now(&self) {
-        self.quit_now.store(true, Ordering::SeqCst);
+        self.quit_now.store(true, AtomicOrdering::SeqCst);
     }
 
     /// Returns true if this worker should quit immediately.
     fn is_quit_now(&self) -> bool {
-        self.quit_now.load(Ordering::SeqCst)
+        self.quit_now.load(AtomicOrdering::SeqCst)
     }
 
     /// Returns the number of pending jobs.
     fn num_pending(&self) -> usize {
-        self.num_pending.load(Ordering::SeqCst)
+        self.num_pending.load(AtomicOrdering::SeqCst)
     }
 
     /// Send work.
     fn send(&self, work: Work) {
-        self.num_pending.fetch_add(1, Ordering::SeqCst);
-        let mut stack = self.stack.lock().unwrap();
-        stack.push(Message::Work(work));
+        self.num_pending.fetch_add(1, AtomicOrdering::SeqCst);
+        self.stack.push(Message::Work(work));
     }
 
     /// Send a quit message.
     fn send_quit(&self) {
-        let mut stack = self.stack.lock().unwrap();
-        stack.push(Message::Quit);
+        self.stack.push(Message::Quit);
     }
 
     /// Receive work.
     fn recv(&self) -> Option<Message> {
-        let mut stack = self.stack.lock().unwrap();
-        stack.pop()
+        self.stack.pop()
     }
 
     /// Signal that work has been finished.
     fn work_done(&self) {
-        self.num_pending.fetch_sub(1, Ordering::SeqCst);
+        self.num_pending.fetch_sub(1, AtomicOrdering::SeqCst);
     }
 }
 
