@@ -1,17 +1,20 @@
-use std::io::{self, Write};
-use std::path::Path;
-use std::time::Instant;
-
-use grep_matcher::{Match, Matcher};
-use grep_searcher::{
-    Searcher, Sink, SinkContext, SinkContextKind, SinkFinish, SinkMatch,
+use std::{
+    io::{self, Write},
+    path::Path,
+    time::Instant,
 };
-use serde_json as json;
 
-use crate::counter::CounterWriter;
-use crate::jsont;
-use crate::stats::Stats;
-use crate::util::find_iter_at_in_context;
+use {
+    grep_matcher::{Match, Matcher},
+    grep_searcher::{
+        Searcher, Sink, SinkContext, SinkContextKind, SinkFinish, SinkMatch,
+    },
+    serde_json as json,
+};
+
+use crate::{
+    counter::CounterWriter, jsont, stats::Stats, util::find_iter_at_in_context,
+};
 
 /// The configuration for the JSON printer.
 ///
@@ -467,7 +470,7 @@ impl<W: io::Write> JSON<W> {
         matcher: M,
     ) -> JSONSink<'static, 's, M, W> {
         JSONSink {
-            matcher: matcher,
+            matcher,
             json: self,
             path: None,
             start_time: Instant::now(),
@@ -493,7 +496,7 @@ impl<W: io::Write> JSON<W> {
         P: ?Sized + AsRef<Path>,
     {
         JSONSink {
-            matcher: matcher,
+            matcher,
             json: self,
             path: Some(path.as_ref()),
             start_time: Instant::now(),
@@ -546,14 +549,13 @@ impl<W> JSON<W> {
 /// This type is generic over a few type parameters:
 ///
 /// * `'p` refers to the lifetime of the file path, if one is provided. When
-///   no file path is given, then this is `'static`.
-/// * `'s` refers to the lifetime of the
-///   [`JSON`](struct.JSON.html)
-///   printer that this type borrows.
+/// no file path is given, then this is `'static`.
+/// * `'s` refers to the lifetime of the [`JSON`] printer that this type
+/// borrows.
 /// * `M` refers to the type of matcher used by
-///   `grep_searcher::Searcher` that is reporting results to this sink.
+/// `grep_searcher::Searcher` that is reporting results to this sink.
 /// * `W` refers to the underlying writer that this printer is writing its
-///   output to.
+/// output to.
 #[derive(Debug)]
 pub struct JSONSink<'p, 's, M: Matcher, W> {
     matcher: M,
